requirements.txt
requests==2.5.3
Flask==0.10.1
Flask-Moment==0.4.0
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.2
MarkupSafe==0.23
praw==3.6.0
pymongo==3.0
python-firebase==1.2
six==1.9.0
update-checker==0.11
Werkzeug==0.10.4
monkeylearn==0.3.5
Flask-PyMongo==0.4.1
